recalc light duration when settings are changed
If the  birthday or  a parameter of the light programm is  chaged the
light  duration is recalculated.

The sunset offset default parameters are changed to a better default
Channel 0
is connected to the ceiling light (A0 on Controllino Maxi Automation)
Channel 1
is connected to the light at the feeding system
Channel 2
is connected to the floor light

The order is chaged so that the chickens are lured to the top.