UpdateLightA0 controls the psu for the leds
The PSU is switched off via a relay when the light is turned off.
In the new coop each light zone is controlled with a seperate LED PSU

This PSU can be controlled via 0-10V Voltage (analog), 10V PWM or a
resistor. In this case the PLC uses a 10V PWM signal that is generated
using a simple voltage divider(2 Resistors) to drop the Voltage from 24V
PWM to 10V PWM.

In the other Coops a big psu is used with seperate dimmers for each
light zone. To save energy each relays can be connected in parallel to
switch the psu one if at least one light requires power.